Store abstraction
The following commit implements a store abstraction around the
underlying data model. This commit introduces mocking via gomock
(see glide update), which allows expectations and understanding
of values from a TDD level.

The store abstraction provides two methods which make it easy
to reason about and the controller (when implemented) doesn't need
to know about a "filesystem" or data store.

 1. Read
 2. Write

This makes changing the store easier in the future, but also has
the massive benefit of being much more testable.