Writing to the store
This commit introduces the ability to write users to the underlying
data store.

This commit also fixes the issue where we were creating
and opening the file in read only mode. Because the fs abstraction
it means that only a very tiny change was required and it that
meant no other code was changed. WIN!